add confirm method to deleteWishList

+++ src/wish_list.js
@@ -87,13 +87,16 @@ class WishList {
     }
 
     deleteWishList() {
+        let confirmDelete = confirm("Are you sure you want to delete? This will remove all items inside your wish list.")
         const wishListId = this.parentElement.dataset.id
-    
-        fetch(`${wishListUrl}/${wishListId}`, {
-            method: "DELETE"
-        })
+        if(confirmDelete) {
+            fetch(`${wishListUrl}/${wishListId}`, {
+                method: "DELETE"
+            })
+            
+            this.parentElement.remove()
+        }
         
-        this.parentElement.remove()
     }
     
 }
